add event loop group
